+++ src/events/client/interactionCreate.js
@@ -38,7 +38,29 @@ module.exports = {
             } catch (error) {
                 console.error(error);
             }
-        } 
+        } else if (interaction.isSelectMenu()) {
+            const { selectMenus } = client;
+            const { customId } = interaction;
+            const menu = selectMenus.get(customId);
+            if(!menu) return new Error('There is no code/file for this select menu. Recommended to create a new one.');
+
+            try {
+                await menu.execute(interaction, client);
+            } catch (error) {
+                console.error(error);
+            }
+        } else if (interaction.type === InteractionType.ModalSubmit) {
+            const { modals } = client;
+            const { customId } = interaction;
+            const modal = modals.get(customId);
+            if(!modal) return new Error('There is no code/file for this modal. Recommended to create a new one.');
+
+            try {
+                await modal.execute(interaction, client);
+            } catch (error) {
+                console.error(error);
+            }
+        }