+++ src/events/client/listener.js
@@ -1,0 +1,10 @@
+module.exports = {
+    name: 'listener',
+    async execute(interaction, client) {
+        let goodbotresponse = ["AntoryBot", "good", "good bot"];
+
+        if(message.content.toLowerCase().includes("good bot")){
+            message.channel.send(goodbotresponse[Math.floor(Math.random() * goodbotresponse.length)]);
+        }
+    }
+}