+++ src/handlers/componentHandler.js
@@ -6,7 +6,7 @@ module.exports = (client) => {
         for(const folder of componentsFolder) {
             const componentFiles = readdirSync(`./src/components/${folder}`).filter(file => file.endsWith('.js'));
 
-            const { buttons } = client;
+            const { buttons, selectMenus, modals } = client;
 
             switch (folder) {
                 case "buttons":
@@ -16,6 +16,19 @@ module.exports = (client) => {
                     }
                     break;
             
+                case "selectMenus":
+                    for(const file of componentFiles) {
+                        const menu = require(`../components/selectMenus/${file}`);
+                        selectMenus.set(menu.data.name, menu);
+                    }
+                    break;
+
+                case "modals":
+                    for(const file of componentFiles) {
+                        const modal = require(`../components/modals/${file}`);
+                        modals.set(modal.data.name, modal);
+                    }
+                    break;
                 default:
                     break;
             }